refactor: 네비게이션 컴포넌트 모듈화 및 SideNav 서버 컴포넌트 전환
